requirements.txt
SQLAlchemy>=0.8.0b1
pyramid
zope.sqlalchemy
transaction
webhelpers
webtest
pyramid_webassets
cssmin
jsmin
